+++ static/app/components/replays/breadcrumbs/breadcrumbDescription.tsx
@@ -22,7 +22,7 @@ interface Props {
   expandPaths?: string[];
 }
 
-export function BreadcrumbItemDescription({
+export function BreadcrumbDescription({
   description,
   allowShowSnippet,
   showSnippet,
